+++ hasklepias-main/examples/ExampleEvents.hs
@@ -40,7 +40,7 @@ type EventData a = (a, a, Text)
 
 toEvent :: (IntervalSizeable a a, Show a) => EventData a -> Event a
 toEvent x = event (beginerval (t1 x) (t2 x))
-                  (context (UnimplementedDomain ()) (packConcepts [t3 x]))
+                  (context (UnimplementedDomain ()) (packConcepts [t3 x]) Nothing)
 
 toEvents :: (Ord a, Show a, IntervalSizeable a a) => [EventData a] -> Events a
 toEvents = sort . map toEvent